Time: 1 ms (99.85%), Space: 41.6 MB (95.2%) - LeetHub

+++ 3379-score-of-a-string/3379-score-of-a-string.java
@@ -3,8 +3,8 @@ class Solution {
         int diff = 0;
         if(s.length()<2) return (int)s.charAt(0);
        for(int i=1; i<s.length(); i++){
-        int count1 = (int)s.charAt(i);
-        int count2 = (int)s.charAt(i-1);
+        int count1 = s.charAt(i);
+        int count2 = s.charAt(i-1);
         diff += Math.abs(count1 - count2);
        } 
        return diff;
